Rename SourcesSimpleView to SourcesTable
