+++ src/test/components/sourcesTable/EmptyStateTable.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Bullseye, Button, EmptyState, EmptyStateBody } from '@patternfly/react-core';
 
 import { componentWrapperIntl } from '../../../utilities/testsHelpers';
-import EmptyStateTable from '../../../components/SourcesSimpleView/EmptyStateTable';
+import EmptyStateTable from '../../../components/SourcesTable/EmptyStateTable';
 import * as actions from '../../../redux/sources/actions';
 
 describe('EmptyStateTable', () => {
